backend/pyproject.toml
[project]
name = "menuwirebackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.1",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "faker>=37.4.2",
    "faker-food>=0.3.0",
    "pillow>=11.3.0",
    "uvicorn[standard]>=0.38.0",
]
